concatenate added headers, rather than appending to array of header elements
This is in line with the HTTP RFC that states, "Multiple message-header fields with the same field-name MAY be present in a message if and only if the entire field-value for that header field is defined as a comma-separated list [i.e., #(values)]. It MUST be possible to combine the multiple header fields into one "field-name: field-value" pair, without changing the semantics of the message, by appending each subsequent field-value to the first, each separated by a comma. The order in which header fields with the same field-name are received is therefore significant to the interpretation of the combined field value, and thus a proxy MUST NOT change the order of these field values when a message is forwarded."

This also adds a getHeader() method, which returns the value for a single header label.

Combined, these make it easier to examine header values, because the value is always a single string, instead of an array of one or more elements that might have to be looped through.

A future addition might be to prevent addHeader() on headers that the RFC specifies have only one entry.